backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
watchdog==3.0.0
psutil==5.9.6
asyncpg==0.29.0
python-dotenv==1.0.0
aiofiles==23.2.1
schedule==1.2.0